feat: Set default language to English
Updates the default language setting from Chinese ('zh') to English ('en') in the application. This includes changes to the `LanguageContext` default and initial state, as well as localized text for the loading indicator and login page. The LanguageSwitcher component has also been removed from the Login page as it is now managed globally.